+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.12.4]
 - Fix `.suggestion` class to ensure it is scrollable and not overlapped by the `.suggestion-actions` fixed element. This change prevents the suggestion content from being hidden.
+- transmit required protocol version to language server
 
 ## [2.12.3]
 - Fix a bug in AI Applyfix on Windows.